fix: explicit reject in execute_rental_request_proposal (#54)
This PR makes the method `execute_rental_request_proposal` reject a
rental request proposal with an explicit reject (reject code 4) and a
textual error message rather than replying with a structured error.

The motivation for this change is that the NNS governance canister
treats NNS functions that reply (with any response, including structured
errors, that the NNS governance canister doesn't even try to parse) as
_executed_. However, we want the NNS function execution to be treated as
_failed_ if the function `execute_rental_request_proposal` returns an
error. This can be achieved by rejecting the call. We go for an explicit
reject (reject code 4) rather than a trap (reject code 5) so that the
history changes are committed to the canister's state (traps roll back
state changes and thus user history won't be updated if the canister
trapped the call).